+++ vertigo-x-notification/src/test/java/io/vertigo/x/notification/NotificationManagerTest.java
@@ -12,13 +12,20 @@ import io.vertigo.x.account.AccountManager;
 import io.vertigo.x.connectors.redis.RedisConnector;
 import io.vertigo.x.notification.data.Accounts;
 
+import java.util.Arrays;
+import java.util.Collection;
+
 import javax.inject.Inject;
 
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
 
+@RunWith(Parameterized.class)
 public class NotificationManagerTest {
 	private App app;
 
@@ -26,20 +33,37 @@ public class NotificationManagerTest {
 	private AccountManager accountManager;
 	@Inject
 	private NotificationManager notificationManager;
-	@Inject
-	private RedisConnector redisConnector;
 
 	private URI<Account> accountURI0;
 	private URI<Account> accountURI1;
 	private URI<Account> accountURI2;
 	private URI<AccountGroup> groupURI;
 
+	@Parameters
+	public static Collection<Object[]> params() {
+		return Arrays.asList(
+				//redis 
+				new Object[] { true },
+				//memory (redis= false)
+				new Object[] { false }
+				);
+	}
+
+	final boolean redis;
+
+	public NotificationManagerTest(final boolean redis) {
+		//params are automatically injected
+		this.redis = redis;
+	}
+
 	@Before
 	public void setUp() {
-		app = new App(MyAppConfig.config());
+		app = new App(MyAppConfig.config(redis));
 		Injector.injectMembers(this, Home.getApp().getComponentSpace());
-		redisConnector.getResource().flushAll();
-
+		if (redis) {
+			final RedisConnector redisConnector = app.getComponentSpace().resolve(RedisConnector.class);
+			redisConnector.getResource().flushAll();
+		}
 		accountURI0 = createAccountURI("0");
 		accountURI1 = createAccountURI("1");
 		accountURI2 = createAccountURI("2");
